backend/requirement.txt
Django
djangorestframework
djangorestframework-simplejwt
transformers
tensorflow
gunicorn
redis
daphne
channels
django-cors-headers
tf-keras